+++ websites/api-docs-smoke-test/gatsby-config.mjs
@@ -15,8 +15,8 @@ const config = {
     title: 'API Docs Smoke Test',
     description: 'Documentation website for API smoke tests',
     breadcrumbs: 'Test Sites',
-    products: [],
-    contentType: null,
+    products: ['smokeTest'],
+    contentType: 'test',
     betaLink: '',
     excludeFromSearchIndex: isProd,
     globalNotification: {